% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{AN.test}
\alias{AN.test}
\title{Adaptive Neyman test in DEGraph package}
\usage{
AN.test(X1, X2, candK = 1:ncol(X1), na.rm = FALSE, pool = FALSE, poolVarX,
  poolVarY)
}
\arguments{
\item{X1}{nxT (number of reps x width of interval). A matrix represents coverage for condition 1.}

\item{X2}{nxT (number of reps x width of interval). A matrix represents coverage for condition 2.}

\item{candK}{a vector, candidate values for the true number of Fourier components.}

\item{na.rm}{A logical value indicating whether variables with NA in at least one of the n1}

\item{poolVarY}{}
}
\value{
The list of adaptive Neyman statistic \code{T}, p.value \code{p}, \code{kstar}, \code{varX}, and  \code{varY}.
}
\description{
Adaptive Neyman test in DEGraph package
}
