% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AN_test}
\alias{AN_test}
\title{Compute the adaptive Neyman test}
\usage{
AN_test(X, Y, na_rm, pool, poolVarX, poolVarY)
}
\arguments{
\item{X}{A coverage vector for first condition}

\item{Y}{A coverage vector for second condition}

\item{na_rm}{TRUE to remove NA values}

\item{pool}{TRUE if variance is pooled. In this case, pre-computed poolVarX and poolVarY must be provided.
If FALSE, pointwise variance is computed.}

\item{poolVarX}{A precomputed variance vector from the first condition. The pooled variance is implemented
according to our paper.}

\item{poolVarY}{A precomputed variance vector from the second condition. The pooled variance is implemented
according to our paper.}
}
\description{
Compute the adaptive Neyman test
}
