% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{aneyman}
\alias{aneyman}
\title{Compute Adaptive Neyman Test after Discrete Fourier Transform.}
\usage{
aneyman(X, Y, maxdim = floor(nrow(X)/2), poolVar = FALSE, numPar = 5,
  candK = 1:nrow(X), na.rm = FALSE)
}
\arguments{
\item{X}{A matrix represents coverage for condition 1 (T x n1).}

\item{Y}{A matrix represents coverage for condition 1 (T x n2).}

\item{maxdim}{the number of the maximum frequency to be tested.
maxdim = L/2 seems a good choice, but you are welcome to use maxdim = L.}

\item{candK}{a vector, candidate values for the true number of components to be tested.}

\item{na.rm}{}
}
\value{
a list of \code{mhat} and Adaptive Neyman test \code{T_AN}.
}
\description{
Compute Adaptive Neyman Test after Discrete Fourier Transform.
}
