% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{bamCoverage}
\alias{bamCoverage}
\title{Count reads mapping to pre-specified regions}
\usage{
bamCoverage(colData, bed_files, mc_cores = 1, sm = 1, binsize = 1)
}
\arguments{
\item{colData}{: a table of sample information}

\item{bed_files}{: full directories to bed files.}

\item{mc_cores}{: number of cores for parallel backend.}

\item{sm}{: smoothing parameter for coverage.}

\item{binsize}{: width of bins to count reads mapping to regions instead of single base pairs.}
}
\value{
: A list of coverage whose elements contain reads from the given regions.
}
\description{
Count reads mapping to pre-specified regions
}
