% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{calculateTotalCounts}
\alias{calculateTotalCounts}
\title{Calculate the matrix of total counts for each genomic regions.}
\usage{
calculateTotalCounts(object, nSamples, bNormWidth, bSampleMean)
}
\arguments{
\item{object}{A \code{tanDb} object}

\item{nSamples}{Sample sizes for each conditions (n1=n2=n)}

\item{bNormWidth}{Parameter for Binning TODO}

\item{bSampleMean}{Parameter for Binning TODO}
}
\value{
the matrix of total counts \code{Ns}
}
\description{
Calculate the matrix of total counts for each genomic regions.
}
