% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{calculateVariance}
\alias{calculateVariance}
\title{Calculuate pooled variance}
\usage{
calculateVariance(object, minus_condition, Global_lower, poolQuant, movAve,
  use_cpp = TRUE, ...)
}
\arguments{
\item{object}{A \code{tanDb} object}

\item{minus_condition}{TRUE if calculating the first (or minus ) condition;
FALSE if calculating the second (or plus) condition.}

\item{Global_lower}{set lower bound for minGlobal (length of pooled var vector for each bins) TODO.
Recommend using \code{Global_lower <- floor(s.size/2)}}

\item{poolQuant}{A pooled quantile at each genomic position.}

\item{movAve}{A number of points for moving average.}

\item{use_cpp}{TRUE if the implementation uses cpp functions;
FALSE if the implementation uses R functions. R version is under development.}
}
\value{
A list of pooled variances for the given condition.
}
\description{
Calculuate pooled variance
}
