% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{calculateWithinSites}
\alias{calculateWithinSites}
\title{Get a list of within-site indices for each bin based
  on total counts \code{Ns} and \code{quantprobs}.}
\usage{
calculateWithinSites(object, quantprobs)
}
\arguments{
\item{object}{A \code{tanDb} object}

\item{quantprobs}{A quantile vector for binning.}
}
\value{
wSites, a list of within-site indices for each bin;
   and dN, a vector of count ranges for each bins.
}
\description{
Get a list of within-site indices for each bin based
  on total counts \code{Ns} and \code{quantprobs}.
}
