% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_Var}
\alias{compute_Var}
\title{Compute the variance vector}
\usage{
compute_Var(X, Y, na_rm, pool)
}
\arguments{
\item{X}{a coverage vector for first condition}

\item{Y}{a coverage vector for second condition}

\item{na_rm}{TRUE to remove NA values}

\item{pool}{TRUE if variance is pooled; FALSE if pointwise variance is computed.
For the current implementation, only case pool = FALSE is considered.}
}
\description{
Compute the variance vector
}
