% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{computePvalues}
\alias{computePvalues}
\title{Compute p values for adaptive tests}
\usage{
computePvalues(object, quant, poolQuant, movAve, Global_lower)
}
\arguments{
\item{object}{A \code{tanDb} object}

\item{quant}{A quantile to obtain the combined p-values.}

\item{poolQuant}{A quantile to pool the variances; DEFAULT is median.}

\item{movAve}{A parameter to smooth the variance.}

\item{Global_lower}{set lower bound for minGlobal (length of pooled var vector for each bins)
Recommend using \code{Global_lower <- floor(s.size/2)}}
}
\description{
Compute p values for adaptive tests
}
