% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_pMat}
\alias{create_pMat}
\title{Title}
\usage{
create_pMat(p.list, totalPeaks, nSamples)
}
\arguments{
\item{p.list}{A list resulted from running \code{computePvals_batch}.}

\item{totalPeaks}{A total number of peaks tested in batch mode.}

\item{nSamples}{The sample size for each conditions (n1=n2=n).}
}
\description{
Title
}
