% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{createDesigns}
\alias{createDesigns}
\title{Create a matrix to store sampling locations for genomic regions.}
\usage{
createDesigns(object, s.size, LHD = TRUE, Uniform = FALSE)
}
\arguments{
\item{object}{A \code{tanDb} object}

\item{s.size}{The number of sampling points allowed in genomic regions.}

\item{LHD}{TRUE if using even spacing for sampling points; always set to TRUE.}

\item{Uniform}{TRUE if allowing uniform point within each grid of continuous sampling points;
always set to FALSE}
}
\value{
Designs, a matrix (total regions x s.size) to store sampling locations for genomic regions
}
\description{
Create a matrix to store sampling locations for genomic regions.
}
