% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{generate_coverage}
\alias{generate_coverage}
\title{Extract reads from a Segvis object}
\usage{
generate_coverage(object, condition, coord = NULL, mc, FUN, ...)
}
\arguments{
\item{object}{A Segvis_block_list object.}

\item{coord}{a vector storing the genomic coordinates to be extracted.}

\item{mc}{number of cores for parallel pipeline.}

\item{...}{}
}
\value{
a data frame contains read coverage.
}
\description{
Extract reads from a Segvis object
}
