% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{generateWithinTan}
\alias{generateWithinTan}
\title{Generate the within (or null) Adaptive Neyman tests}
\usage{
generateWithinTan(object, minus_condition)
}
\arguments{
\item{object}{A \code{tanDb} object}

\item{minus_condition}{TRUE if calculating the first (or minus ) condition;
FALSE if calculating the second (or plus) condition.}
}
\description{
Generate the within (or null) Adaptive Neyman tests
}
