% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{getXY}
\alias{getXY}
\title{Give the tag Counts and genomic positions of a given consensus interval.}
\usage{
getXY(site, Rep, c_ranges, normalized = 1)
}
\arguments{
\item{site}{Site indexed in c_ranges.}

\item{Rep}{Replication ChIP data which is obtained from loading RData.}

\item{c_ranges}{Ranges(GRanges object).}

\item{normalized}{A vector of normalized values for each ChIP samples.}
}
\value{
coord (Genomic Positions) and y.coord (normalized tag Counts)
}
\description{
Give the tag Counts and genomic positions of a given consensus interval.
}
\author{
Duy Nguyen on April 6, 2015
}
