% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{mmd}
\alias{mmd}
\title{Compute the MMD distance in MMDiff package}
\usage{
mmd(X, labels, sigma = NULL)
}
\arguments{
\item{X}{An 1-by-T matrix to store 2 curves}

\item{labels}{1 for first curve; -1 for second curve}

\item{sigma}{NULL}
}
\value{
A data frame of MMD; MMD distance is stored in d$val
}
\description{
Compute the MMD distance in MMDiff package
}
