% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{movingAverage}
\alias{movingAverage}
\title{Moving Average}
\usage{
movingAverage(x, n = 1, centered = FALSE)
}
\arguments{
\item{x}{the vector}

\item{n}{the number of samples}

\item{centered}{if FALSE, then average current sample and previous (n-1) samples}
}
\description{
Moving Average
}
