% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plotCoverage}
\alias{plotCoverage}
\title{create a coverage profile plot}
\usage{
plotCoverage(coverage, title = "", size = 0.5, showLegend = TRUE)
}
\arguments{
\item{coverage}{: A matrix to store coverage reads}

\item{title}{: title of the plot}

\item{size}{: thickness of lines}

\item{showLegend}{}
}
\value{
A ggplot2 plot
}
\description{
create a coverage profile plot
}
