% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_designs.R
\name{Sampling}
\alias{Sampling}
\title{Perform Latin Hypercube sampling or uniform sampling on a genomic interval.}
\usage{
Sampling(s.size, start = 1, end = 1000, LHD = TRUE, Uniform = TRUE)
}
\arguments{
\item{s.size}{size of sampling points.}

\item{start}{start of the genomic interval.}

\item{end}{end of the genomic interval.}

\item{LHD}{True if performing LH sampling; False if performing uniform sampling.}

\item{Uniform}{for case LHD = TRUE, if Uniform = TRUE, we
sample points uniformly for each bin; FALSE if sampling points are end points.}
}
\value{
An one dimenstional array storing the sampling genomic positions
}
\description{
Perform Latin Hypercube sampling or uniform sampling on a genomic interval.
}
\author{
Duy Nguyen on April 6, 2015; modified on February 24, 2016
}
