% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{tanDb-class}
\alias{tanDb-class}
\title{A class to store ChIP data and all related objects for differential analysis.}
\description{
A class to store ChIP data and all related objects for differential analysis.
}
\section{Slots}{

\describe{
\item{\code{coverage}}{A list stores all coverage data. The length of this list
is equal to the number of pre-defined genomic regions to be tested.}

\item{\code{Designs}}{A matrix stores the sampling locations for genomic regions.}

\item{\code{Ns}}{A matrix stores the total counts for each genomic regions and
across library}

\item{\code{s.size}}{The number of sampling points allowed in genomic regions.}

\item{\code{wSites}}{A list of within-site indices for each bin based
on total counts \code{Ns} and \code{quantprobs}.}

\item{\code{quantprobs}}{A quantile vector for binning.}

\item{\code{nSamples}}{The sample size for each conditions (n1=n2=n)}

\item{\code{dN}}{a vector of count ranges for each bins based on \code{quantprobs}.}

\item{\code{minusVar}}{A list of pooled variances for the first condition. Pooling method for
variance is disscussed in our paper.}

\item{\code{plusVar}}{A list of pooled variances for the second condition. Pooling method for
variance is disscussed in our paper.}

\item{\code{poolVar}}{A list of pooled variances for case sanple size \code{n = 2}.}

\item{\code{W1}}{A matrix stores within Adaptive Neyman tests for first condition.}

\item{\code{W2}}{A matrix stores within Adaptive Neyman tests for second condition.}

\item{\code{W}}{A matrix stores Adaptive Neyman tests under the null hypothesis. This is only for case sample size \code{n=2}.}

\item{\code{PvalList}}{A list of results (\code{('pval','FDR')})for P-values.}

\item{\code{pMat}}{A matrix with number of rows equal to number of peaks \code{x}. \code{pMat} stores all the p-values
from running computePvalues in batch mode (useful when testing very large number of peaks).}

\item{\code{p.list}}{A list to store results in data frame format when running batch mode.}

\item{\code{binsCompleted}}{A vector stores all bins that are succesfullly run in batch mode.}

\item{\code{sitesUnused}}{A vector of indices for unused sites due to bad quality peaks during testing within and between TAN.}
}}

