% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{computePvalues}
\alias{computePvalues}
\title{Compute p values for adaptive tests}
\usage{
computePvalues(object, quant, poolQuant, movAve)
}
\arguments{
\item{object}{A \code{tanDb} object}

\item{quant}{A quantile to obtain the combined p-values.}

\item{poolQuant}{A quantile to pool the variances; DEFAULT is median.}

\item{movAve}{A parameter to smooth the variance.}
}
\description{
Compute p values for adaptive tests
}
