% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{computePvalues}
\alias{computePvalues}
\title{Compute p values for adaptive tests}
\usage{
computePvalues(object, quant, poolQuant, movAve, Global_lower, use_cpp = TRUE,
  ignore_sitesUnused = TRUE, na_impute = FALSE, ...)
}
\arguments{
\item{object}{A \code{tanDb} object}

\item{quant}{A quantile to obtain the combined p-values.}

\item{poolQuant}{A quantile to pool the variances; DEFAULT is median.}

\item{movAve}{A parameter to smooth the variance.}

\item{Global_lower}{set lower bound for minGlobal (length of pooled var vector for each bins)
Recommend using \code{Global_lower <- floor(s.size/2)}}

\item{use_cpp}{TRUE if the implementation uses cpp functions;
FALSE if the implementation uses R functions.}

\item{ignore_sitesUnused}{FALSE if considering sitesUnused when computing p-values of testing between;
TRUE by default since sitesUnused caused many zero p-values due to low-count/bad regions in sitesUnused.}

\item{na_impute}{Impute missing value of \code{p-values} and \code{FDR} matrices by taking global mean for all p;FALSE by default.}
}
\description{
Compute p values for adaptive tests
}
