% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{computePvalues}
\alias{computePvalues}
\title{Compute p values for adaptive tests}
\usage{
computePvalues(object, quant, poolQuant, movAve, Global_lower, use_cpp = TRUE,
  ignore_sitesUnused = FALSE, na_impute = TRUE, ...)
}
\arguments{
\item{object}{A \code{tanDb} object}

\item{quant}{A quantile to obtain the combined p-values.}

\item{poolQuant}{A quantile to pool the variances; DEFAULT is median.}

\item{movAve}{A parameter to smooth the variance.}

\item{Global_lower}{set lower bound for minGlobal (length of pooled var vector for each bins)
Recommend using \code{Global_lower <- floor(s.size/2)}}

\item{use_cpp}{TRUE if the implementation uses cpp functions;
FALSE if the implementation uses R functions. R version is under development.}

\item{ignore_sitesUnused}{FALSE by default if considering sitesUnused when computing p-values of testing between;
recommend using TRUE since sitesUnused caused pvalues = 0. #TODO: clean up}

\item{na_impute}{Impute missing value of \code{p-values} and \code{FDR} matrices by taking global mean for all p;TRUE by default. Recommend using TRUE}
}
\description{
Compute p values for adaptive tests
}
