% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{evalPvals}
\alias{evalPvals}
\title{Eval p-values from different quantile.}
\usage{
evalPvals(P, total = nrow(P[["pval"]]), quant = 1, nSamples, BH = FALSE,
  na.rm = TRUE)
}
\arguments{
\item{P}{A matrix of P values whose columns a testing between samples across conditions.}

\item{total}{Total genomic intervals.}

\item{quant}{Quantile used for combined pvales.}

\item{nSamples}{The sample size for each conditions (n1=n2=n).}

\item{BH}{A logical if multiple testing correction using B-H method is applied. Defaul is FALSE.}

\item{na.rm}{logical; if true, any NA and NaN's are removed before computing.}
}
\value{
Combined (Raw) p-values with the given quantile.
}
\description{
Eval p-values from different quantile.
}
\author{
Duy Nguyen on February 26, 2016
}
