% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{evalPvals}
\alias{evalPvals}
\title{Eval p-values from different quantile.}
\usage{
evalPvals(P, total = nrow(P[["pval"]]), quant = 1, Between_cols)
}
\arguments{
\item{P}{A matrix of P values whose columns a testing between samples across conditions.}

\item{total}{Total genomic intervals.}

\item{quant}{Quantile used for combined pvales.}

\item{Between_cols}{A number of columns for testing between conditions.
(e.g., n = 4, there are 6*6 columns)}
}
\value{
Combined p-values with the given quantile.
}
\description{
Eval p-values from different quantile.
}
\author{
Duy Nguyen on February 26, 2016
}
